@Innixma 2023-02-14
Benchmark result looks good! Pareto dominates mainline (2022_02_12 is mainline, 2022_02_14 is this PR): 

<img width="1211" alt="autogluon-benchmark_–_run_check_inference_speed_py__autogluon-utils_" src="https://user-images.githubusercontent.com/16392542/218807349-93e32023-46ff-4894-ac2e-80d70da44b94.png">